+++ plugins/org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/artifact/PurgeArtifacts.java
@@ -96,7 +96,7 @@ public class PurgeArtifacts extends AbstractDbTxOperation {
          }
          artJoin2.store();
 
-         int queryId = artJoin2.getQueryId();
+         Long queryId = artJoin2.getQueryId();
 
          try (Id4JoinQuery txJoin = JoinUtility.createId4JoinQuery(getJdbcClient(), connection)) {
             insertSelectItems(txJoin, connection, "osee_relation_link",
@@ -143,7 +143,7 @@ public class PurgeArtifacts extends AbstractDbTxOperation {
       }
    }
 
-   public void insertSelectItems(Id4JoinQuery txJoin, JdbcConnection connection, String tableName, String artifactJoinSql, int queryId) {
+   public void insertSelectItems(Id4JoinQuery txJoin, JdbcConnection connection, String tableName, String artifactJoinSql, Long queryId) {
       String query = String.format(SELECT_ITEM_GAMMAS, getJdbcClient().getOrderedHint(), tableName, artifactJoinSql);
       try (JdbcStatement chStmt = getJdbcClient().getStatement(connection)) {
          chStmt.runPreparedQuery(query, queryId);
