+++ plugins/org.eclipse.osee.orcs.db/src/org/eclipse/osee/orcs/db/internal/search/indexer/IndexerCallableFactory.java
@@ -23,6 +23,6 @@ import org.eclipse.osee.orcs.search.IndexerCollector;
  */
 public interface IndexerCallableFactory {
 
-   Callable<?> createIndexerTaskCallable(OrcsSession session, OrcsTokenService tokenService, IndexerCollector collector, int queryId);
+   Callable<?> createIndexerTaskCallable(OrcsSession session, OrcsTokenService tokenService, IndexerCollector collector, Long queryId);
 
 }
